Reverted resolveAsset and resolveContract to use Hardhat Moved registerValidators and stakeValidators into utils without a Hardhat dependency Revert getBlock to use Hardhat for the provider
